requirements.txt
streamlit
pandas
openpyxl
psycopg2-binary
python-dotenv